// flask-q4-recommendations-demo/new-q4-app/src/API.ts
export enum AnnotationTaskStatus {
  TODO = "TODO",
  FAILED = "FAILED",
  DONE = "DONE",
}


export type NerAnnotationSampleInput = {
  text: string,
  annotations?: Array< NerAnnotationInput | null > | null,
};

export type NerAnnotationInput = {
  text: string,
  tag?: AnnotationTagInput | null,
  start: number,
  end: number,
};

export type AnnotationTagInput = {
  id: string,
  name: string,
  description: string,
};

export type NerAnnotationTask = {
  __typename: "NerAnnotationTask",
  id: string,
  project_id: string,
  instruction: AnnotationTaskInstruction,
  status: AnnotationTaskStatus,
  type: AnnotationTaskType,
  metadata?: string | null,
  annotator_id: string,
  tags:  Array<AnnotationTag >,
  sample: NerAnnotationSample,
  issue?: AnnotationTaskIssue | null,
  createdAt: string,
  updatedAt: string,
};
// ...
export type AnnotationTaskInstruction = {
  __typename: "AnnotationTaskInstruction",
  text: string,
  full_instructions_url?: string | null,
};

export enum AnnotationTaskType {
  NER = "NER",
  CLASSIFICATION = "CLASSIFICATION",
  ALLERGIES = "ALLERGIES",
}


export type AnnotationTag = {
  __typename: "AnnotationTag",
  id: string,
  name: string,
  description: string,
};

export type ClassificationAnnotationSample = {
  __typename: "ClassificationAnnotationSample",
  text: string,
  annotations:  Array<ClassificationAnnotation | null >,
};

export type ClassificationAnnotation = {
  __typename: "ClassificationAnnotation",
  text: string,
  tag?: AnnotationTag | null,
};

export type AnnotationTaskIssue = {
  __typename: "AnnotationTaskIssue",
  id: string,
  type: AnnotationTaskIssueType,
  reason?: AnnotationTaskIssueReason | null,
  notes: string,
  createdAt: string,
  updatedAt: string,
};

export type AnnotationTaskIssueType = {
  __typename: "AnnotationTaskIssueType",
  id: string,
  type: string,
  createdAt: string,
  updatedAt: string,
};

export type AnnotationTaskIssueReason = {
  __typename: "AnnotationTaskIssueReason",
  id: string,
  reason: string,
  createdAt: string,
  updatedAt: string,
};
// ...
export type NerAnnotationSample = {
  __typename: "NerAnnotationSample",
  text: string,
  annotations?:  Array<NerAnnotation | null > | null,
};

export type NerAnnotation = {
  __typename: "NerAnnotation",
  text: string,
  tag?: AnnotationTag | null,
  start: number,
  end: number,
};

export type CreateAnnotatorInput = {
  id?: string | null,
  username: string,
  name: string,
  email: string,
};